correct number of functions in files
